added test cases and modified set_raid_mode and set_failed_disk functions to make parity blocks and restore the disks

+++ kernel/proc.h
@@ -124,9 +124,9 @@ struct proc
   int pages_swapped_in;     // Number of pages swapped in
   int pages_swapped_out;    // Number of pages swapped out (same as evicted but for swapped processes)
   int resident_pages;       // Number of resident pages
-  int disk_reads;           // Number of disk reads
-  int disk_writes;          // Number of disk writes
-  int avg_disk_latency;   // Average disk latency
+  uint64 disk_reads;           // Number of disk reads
+  uint64 disk_writes;          // Number of disk writes
+  uint64 avg_disk_latency;   // Average disk latency
   
   // wait_lock must be held when using this:
   struct proc *parent; // Parent process
